import tests for OpenVPN plugin

+++ lib/openvpn-auth-oauth2/openvpn/plugin.go
@@ -168,7 +168,7 @@ func PluginClientConstructorV1(handlePtr c.OpenVPNPluginHandle) c.OpenVPNPluginC
 	clientHandle := cgo.NewHandle(&ClientContext{})
 
 	//goland:noinspection GoVetUnsafePointer
-	return unsafe.Pointer(clientHandle)
+	return unsafe.Pointer(uintptr(clientHandle))
 }
 
 func PluginClientDestructorV1(handlePtr c.OpenVPNPluginHandle, perClientContext c.OpenVPNPluginClientContext) {
